Methods to estimate solvent's T-dep densities added
A json file containing solvents' critical temperature, critical
density, and coolprop name has been added.

And methods to estimate solvent's temperature-dependent liquid phase and
gas phase saturation densities are added. If the temperature is out of
valid range, it also provides the error message with the valid
temperature range. If `T_adjustment` is set to True, then the input
temperature is automatically adjusted to the closest valid temperature
when it is outside the valid range.