Add an option to provide solvent critical property dictionary
An option to provide `solv_crit_prop_dict` to class `SolubilityCalculations`
is added. This optional allows one to provide their own dictionary
containing solvent's critical temperature, critical density, and
CoolProp name. These solvent data are required for T-dep solubility
calculation.

If this option is not used, then the default `solv_crit_prop_dict` is
loaded. This option is added so more solvents can be added to the
dictionary in the future.